Finished database and tested for good values.  Everything is working just dandy.  after lunch i'm hoping to give it some shit values and make sure it doesn't break :)

+++ SolarPowerCalculatorProject/src/project/server/SunData.java
@@ -10,7 +10,7 @@ public class SunData {
 	//out how to calculate using solarExposure
 	private double solarExposure;//kwH/m^2
 	private double sunlightHours;
-	private int latitude;
+	private double latitude;
 	//maybe add monthly stats later
 	
 	//default
